requirements.txt
faster-whisper==1.2.1
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
